+++ ryft-ui/src/main/java/com/ryftpay/android/ui/model/googlepay/RyftTokenizationSpecification.kt
@@ -1,4 +1,4 @@
-package com.ryftpay.android.ui.service.request.googlepay
+package com.ryftpay.android.ui.model.googlepay
 
 import com.ryftpay.android.core.model.api.RyftPublicApiKey
