+++ ryft-ui/src/main/java/com/ryftpay/android/ui/service/DefaultGooglePayService.kt
@@ -4,8 +4,8 @@ import android.app.Activity
 import com.google.android.gms.tasks.Task
 import com.google.android.gms.wallet.AutoResolveHelper
 import com.google.android.gms.wallet.PaymentsClient
-import com.ryftpay.android.ui.service.request.googlepay.GooglePayRequestFactory
-import com.ryftpay.android.ui.service.request.googlepay.LoadPaymentDataRequest
+import com.ryftpay.android.ui.model.googlepay.LoadPaymentDataRequest
+import com.ryftpay.android.ui.service.request.GooglePayRequestFactory
 
 internal class DefaultGooglePayService(
     private val paymentsClient: PaymentsClient
